+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    interoperable with Git packfile index files.
    ([#12](https://github.com/gg-scm/gg-git/issues/12))
 -  `packfile.ReadHeader` enables random access to a packfile.
+-  `object.Prefix` allows marshaling and unmarshaling the `"blob 42\x00"` prefix
+   used as part of the Git object hash.
 
 ### Changed